vehicle_reorg: split vehicle part use into its own file
vehicle parts usage needs to be even more substantially refactored, but
step one is putting everything related to use a vehicle part in its own
file.
